Move bootstrap test to Swift Testing file
